First commit : data preprocessing - Convert Categorical columns to numeric codes
When you’ve been devastated by a serious car accident, your focus is on
the things that matter the most: family, friends, and other loved ones.
Pushing paper with your insurance agent is the last place you want your
time or mental energy spent. Allstate is currently developing automated
methods of predicting the cost, and hence severity, of claims. In this
recruitment challenge,